+++ csrc/layernorm.cpp
