+++ csrc/utils.h
